refactor: Return default value if Secret Manager client is None
This change refactors the `get_secret_key` method in `SecretKeyGoogleCloud` to return the `default_value` if the `client` attribute is None.

This prevents errors when the Secret Manager client is not initialized, for example when running locally without Google Cloud credentials.  Without this change, the code would attempt to access the `client` attribute, resulting in an AttributeError.