refactor: Make default_value a required parameter for get_secret_key
This change modifies the `get_secret_key` method in the `SecretKeyBase` class to make the `default_value` parameter required. It removes the `Optional` type hint from the `default_value` parameter and updates the method implementations in `MockSecretKey` and `SecretKeyGoogleCloud` accordingly.

The `MockSecretKey` default value is now "mock_value" if not given.