refactor: Simplify ErrorWithGoogleCloudAuthentication
The ErrorWithGoogleCloudAuthentication class is refactored to take a simple message string in the constructor instead of a GoogleAuthError object. This simplifies the instantiation of the exception and reduces dependencies on Google Cloud libraries. The original implementation included formatting within the exception class, moving the responsiblity of exception message construction to the caller.