refactor: Rename DatabaseSettings attributes to be private
This change renames the attributes of the `DatabaseSettings` class to be private by prefixing them with an underscore. This is done to better encapsulate the internal state of the class and prevent direct access from outside. Access to these parameters should be done through public methods (e.g. properties). There are no breaking changes.