feat: Add support for SQLite databases
This commit adds explicit support for SQLite databases using `aiosqlite`. This is done by checking if the database URL starts with "sqlite+aiosqlite://". If it does, the `create_async_engine` function is called with the database URL.

This change is necessary to ensure that SQLite databases are correctly configured when running tests.